added method implementations to contribution, compensation, and productivity

+++ src/EmployeeType.java
@@ -2,5 +2,5 @@
 public interface EmployeeType {
 	public double calcContribution(Employee e);
 	public double calcCompensation(Employee e);
-	public double analyzeProductivity(double cont, double comp);
+	public int analyzeProductivity(double cont, double comp);
 }
